feat: Setup Homebrew and Cargo with install scripts
This commit introduces scripts to setup Homebrew and Cargo, along with their
respective packages, during the installation process.

The install.conf.yaml file is updated to include default linking settings
and to execute the new setup scripts.

The setup-homebrew.sh script installs Homebrew and a list of essential
packages.

The setup-cargo.sh script installs Cargo and a list of rust-based tools.